requirements.txt
setuptools
boto3
Werkzeug==2.2.3 #3.1.3
flask==2.2.5
python-dotenv
pytz
pillow



